+++ libs/idp-server-platform/src/main/java/org/idp/server/platform/security/repository/SecurityEventQueryRepository.java
@@ -26,7 +26,7 @@ import java.util.List;
 
 public interface SecurityEventQueryRepository {
 
-  int findTotalCount(Tenant tenant, SecurityEventQueries queries);
+  long findTotalCount(Tenant tenant, SecurityEventQueries queries);
 
   List<SecurityEvent> findList(Tenant tenant, SecurityEventQueries queries);
